refactor: Extract FDv1 fallback directive parsing into source module
