+++ launchdarkly-server-sdk/src/fdv2/source.rs
@@ -1,10 +1,15 @@
 use std::time::Duration;
 
 use futures::future::BoxFuture;
+use http::HeaderMap;
 
 use crate::stores::change_set::ChangeSet;
 
 use super::model::Selector;
+
+pub(super) const FALLBACK_HEADER: &str = "X-LD-FD-Fallback";
+pub(super) const FALLBACK_TTL_HEADER: &str = "X-LD-FD-Fallback-TTL";
+pub(super) const DEFAULT_FALLBACK_TTL: Duration = Duration::from_secs(60 * 60);
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) enum ErrorKind {
@@ -24,6 +29,20 @@ pub(crate) struct ErrorInfo {
 #[derive(Debug)]
 pub(crate) struct FDv1FallbackDirective {
     pub(crate) ttl: Duration,
+}
+
+pub(super) fn read_fallback_directive(headers: &HeaderMap) -> Option<FDv1FallbackDirective> {
+    let flag = headers.get(FALLBACK_HEADER)?;
+    if !flag.to_str().is_ok_and(|s| s.eq_ignore_ascii_case("true")) {
+        return None;
+    }
+    let ttl = headers
+        .get(FALLBACK_TTL_HEADER)
+        .and_then(|v| v.to_str().ok())
+        .and_then(|s| s.parse::<u64>().ok())
+        .map(Duration::from_secs)
+        .unwrap_or(DEFAULT_FALLBACK_TTL);
+    Some(FDv1FallbackDirective { ttl })
 }
 
 #[derive(Debug)]
@@ -49,4 +68,55 @@ pub(crate) trait Initializer: Send {
 pub(crate) trait Synchronizer: Send {
     fn next(&mut self, selector: Selector) -> BoxFuture<'_, FDv2SourceEvent>;
     fn name(&self) -> &str;
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn headers(pairs: &[(&str, &str)]) -> HeaderMap {
+        let mut h = HeaderMap::new();
+        for (k, v) in pairs {
+            h.insert(
+                http::HeaderName::from_bytes(k.as_bytes()).unwrap(),
+                http::HeaderValue::from_str(v).unwrap(),
+            );
+        }
+        h
+    }
+
+    #[test]
+    fn fallback_absent_header_returns_none() {
+        assert!(read_fallback_directive(&HeaderMap::new()).is_none());
+    }
+
+    #[test]
+    fn fallback_header_value_other_than_true_returns_none() {
+        let h = headers(&[("X-LD-FD-Fallback", "false")]);
+        assert!(read_fallback_directive(&h).is_none());
+    }
+
+    #[test]
+    fn fallback_header_uppercase_true_uses_default_ttl() {
+        let h = headers(&[("X-LD-FD-Fallback", "TRUE")]);
+        let d = read_fallback_directive(&h).expect("directive");
+        assert_eq!(d.ttl, DEFAULT_FALLBACK_TTL);
+    }
+
+    #[test]
+    fn fallback_ttl_header_is_parsed() {
+        let h = headers(&[("X-LD-FD-Fallback", "true"), ("X-LD-FD-Fallback-TTL", "60")]);
+        let d = read_fallback_directive(&h).expect("directive");
+        assert_eq!(d.ttl, Duration::from_secs(60));
+    }
+
+    #[test]
+    fn fallback_ttl_header_malformed_uses_default() {
+        let h = headers(&[
+            ("X-LD-FD-Fallback", "true"),
+            ("X-LD-FD-Fallback-TTL", "not-a-number"),
+        ]);
+        let d = read_fallback_directive(&h).expect("directive");
+        assert_eq!(d.ttl, DEFAULT_FALLBACK_TTL);
+    }
 }
